# optimized LRU (70% less cpu time)