--- a/lib/mini_sql/postgres/deserializer_cache.rb
+++ b/lib/mini_sql/postgres/deserializer_cache.rb
@@ -14,15 +14,10 @@ module MiniSql
       def materializer(result)
         key = result.fields
 
-        materializer = @cache.delete(key)
-        if materializer
-          @cache[key] = materializer
-        else
-          materializer = @cache[key] = new_row_matrializer(result)
-          @cache.shift if @cache.length > @max_size
+        @cache[key] || begin
+            @cache.shift if @cache.size + 1 > @max_size
+            @cache[key] = new_row_matrializer(result)
         end
-
-        materializer
       end
 
       def materialize(result, decorator_module = nil)
@@ -31,13 +26,11 @@ module MiniSql
         key = result.fields
 
         # trivial fast LRU implementation
-        materializer = @cache.delete(key)
-        if materializer
-          @cache[key] = materializer
-        else
-          materializer = @cache[key] = new_row_matrializer(result)
-          @cache.shift if @cache.length > @max_size
-        end
+        materializer =
+          @cache[key] || begin
+            @cache.shift if @cache.size + 1 > @max_size
+            @cache[key] = new_row_matrializer(result)
+          end
 
         materializer.include(decorator_module) if decorator_module